Start on implimenting the Minecraft protocol.
Right now just handles the login event and sending the motd / servername.
This is the structure I feel we can use to process the rest of the packets.
It also leaves room for us to switch to processing multiple packets per tcp message,
if we discover clients which do that.