Process all client packets
More code is needed in the server to provide / track things like player
id and spawn position, but all of the packet the client can send to the server
are handled.

I've also added in a way for the game server to send packets back to the client
by passing the tuples that the builder methods take.